Unify software/dataset CSV sync into scripts/sync_contributions.py
Renames scripts/sync_software_contributions.py to sync_contributions.py
and extends it to also handle direct Datasets-branch form responses,
so both Contributed Software and Contributed Datasets cards are kept
in sync from one script and one pair of CSV inputs.

- Rows where the form's "Contribution deliverable type" is Datasets now
  create/update docs/contribution-types/_data/datasets/<id>.yaml
  directly, reusing the same field mapping already used for a Software
  submission's associated_dataset sub-block (plus that submission's own
  UAT answers, which the associated_dataset block intentionally omits).
- New review gate: an existing card's form_data is never overwritten
  silently. Any change -- a resubmission, or a team editing their
  response in place without resubmitting -- prints a before/after diff
  per field and prompts to apply or skip (y/n/all/quit). --yes bypasses
  this for a bulk sync. Brand-new cards are still created outright and
  flagged needs_review, since there's nothing to overwrite.
- An existing Datasets record's title/country/institute are now
  protected once set (only assigned at creation), since the current 17
  records were hand-polished during the original page backfill rather
  than sourced from the spreadsheet. Software records are unaffected
  and keep refreshing those fields from the spreadsheet every run.
- Fixed a bug (caught via synthetic-CSV testing, not yet exercised
  against real data): the software sync loop iterated every non-General
  Pool spreadsheet row, so once the spreadsheet started carrying
  Datasets-category rows too, they were being swept in and creating
  bogus software cards. Now explicitly scoped to the known software
  category codes (4.2/4.3).

conf.py's loader-docstring comments updated to point at the new
filename. Not wired into CI/tox -- same manual, locally-run design as
before.
