+++ conf.py
@@ -39,10 +39,11 @@ linkcheck_ignore = [
 #
 # Each dataset record lives as a YAML file in
 # docs/contribution-types/_data/datasets/<contribution-id>.yaml, split into
-# a `form_data` section (owned by the CSV pull script; safe to overwrite)
-# and a `curated` section (hand-edited only; the pull script never touches
-# it). This loader reads all records, resolves each one's display status,
-# and exposes them to the contributed-datasets.rst page via sphinx_jinja.
+# a `form_data` section (owned by scripts/sync_contributions.py; safe to
+# overwrite) and a `curated` section (hand-edited only; the sync script
+# never touches it). This loader reads all records, resolves each one's
+# display status, and exposes them to the contributed-datasets.rst page via
+# sphinx_jinja.
 # ============================================================================
 
 def _slugify(value):
@@ -562,9 +563,10 @@ jinja_contexts["contributed_opportunities"]["cid_to_slugs"] = (
 #
 # Each software-contribution record lives as a YAML file in
 # docs/contribution-types/_data/software/<contribution-id>.yaml, split into
-# a `form_data` section (owned by scripts/sync_software_contributions.py;
-# safe to overwrite) and a `curated` section (hand-edited only; the sync
-# script never touches it) -- same split as the Contributed Datasets page.
+# a `form_data` section (owned by scripts/sync_contributions.py; safe to
+# overwrite) and a `curated` section (hand-edited only; the sync script
+# never touches it) -- same split as the Contributed Datasets page. Both
+# pages are synced by the same script; see its module docstring.
 # General Pool contributions are not part of this dataset at all (they have
 # their own page); every record here is Directable or Non-directable.
 # ============================================================================
